RPLWinlineBot/app/requirements.txt
anyio==4.6.0
certifi==2024.8.30
colorama==0.4.6
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
numpy==2.1.1
openpyxl==3.1.5
pandas==2.2.3
psycopg==3.2.3
psycopg-binary==3.2.3
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
pytz==2024.2
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.2
